# src/WebUI/Server/wwwroot/api/v1/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CleanArchitecture API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Auctions/{userId}/won": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_GetWonAuctionsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionItemsVM"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{userId}/created": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_GetCreatedAuctionsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionItemsVM"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{id}/view": {
      "get": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_GetAuction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionItemDetailsVM"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/new": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_AddAuction",
        "requestBody": {
          "x-name": "newAuctionItem",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionItemDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{id}": {
      "put": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_PutAuctionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updatedAuctionItem",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionItemDTO"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_DeleteAuctionItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{auctionItemId}/buy/{userId}": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_BuyAuctionItem",
        "parameters": [
          {
            "name": "auctionItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{userId}/validate/{auctionItemId}": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_ValidateAuctionItem",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "auctionItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auctions/{auctionItemId}/bid": {
      "post": {
        "tags": [
          "Auctions"
        ],
        "operationId": "Auctions_PlaceAuctionBid",
        "parameters": [
          {
            "name": "auctionItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "bid",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidDTO"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuctionItemsVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionItemDTO"
            }
          }
        }
      },
      "AuctionItemDTO": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "description",
          "images",
          "startDate",
          "endDate",
          "category",
          "startingBidAmount",
          "buyItNowPrice",
          "reservePrice",
          "deliveryMethod"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "publicId": {
            "type": "string",
            "format": "guid"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "category": {
            "minimum": 1.0,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "startingBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          },
          "currentBidAmount": {
            "type": "number",
            "format": "decimal"
          },
          "buyItNowPrice": {
            "type": "number",
            "format": "decimal"
          },
          "reservePrice": {
            "type": "number",
            "format": "decimal"
          },
          "deliveryMethod": {
            "minimum": 1.0,
            "oneOf": [
              {
                "$ref": "#/components/schemas/DeliveryMethod"
              }
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "publicStatus": {
            "$ref": "#/components/schemas/PublicStatus"
          },
          "biddingHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BidDTO"
            }
          },
          "lat": {
            "type": "number",
            "format": "float",
            "maximum": 85.0,
            "minimum": -85.0
          },
          "lon": {
            "type": "number",
            "format": "float",
            "maximum": 180.0,
            "minimum": -180.0
          },
          "winningBidder": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Category": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "None",
          "Books",
          "Music",
          "Movies",
          "Beauty",
          "Clothing",
          "Accessories",
          "Electronics",
          "Vehicles",
          "Toys",
          "Sport",
          "HomeAndGardening",
          "CollectiblesAndMemorabilia"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ]
      },
      "DeliveryMethod": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "None",
          "DeliveryByCourier",
          "PersonalPickUp"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "PublicStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "None",
          "Active",
          "Closed"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "BidDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "auctionItemId": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AuctionItemDetailsVM": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auction": {
            "$ref": "#/components/schemas/AuctionItemDTO"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}